+++ crates/swc/src/lib.rs
@@ -458,103 +458,6 @@ impl Compiler {
     }
 }
 
-struct SwcSourceMapConfig<'a> {
-    source_file_name: Option<&'a str>,
-    /// Output path of the `.map` file.
-    output_path: Option<&'a Path>,
-
-    names: &'a AHashMap<BytePos, JsWord>,
-
-    inline_sources_content: bool,
-
-    emit_columns: bool,
-}
-
-impl SourceMapGenConfig for SwcSourceMapConfig<'_> {
-    fn file_name_to_source(&self, f: &FileName) -> String {
-        if let Some(file_name) = self.source_file_name {
-            return file_name.to_string();
-        }
-
-        let base_path = match self.output_path {
-            Some(v) => v,
-            None => return f.to_string(),
-        };
-        let target = match f {
-            FileName::Real(v) => v,
-            _ => return f.to_string(),
-        };
-
-        let rel = pathdiff::diff_paths(target, base_path);
-        match rel {
-            Some(v) => {
-                let s = v.to_string_lossy().to_string();
-                if cfg!(target_os = "windows") {
-                    s.replace('\\', "/")
-                } else {
-                    s
-                }
-            }
-            None => f.to_string(),
-        }
-    }
-
-    fn name_for_bytepos(&self, pos: BytePos) -> Option<&str> {
-        self.names.get(&pos).map(|v| &**v)
-    }
-
-    fn inline_sources_content(&self, _: &FileName) -> bool {
-        self.inline_sources_content
-    }
-
-    fn emit_columns(&self, _f: &FileName) -> bool {
-        self.emit_columns
-    }
-
-    fn skip(&self, f: &FileName) -> bool {
-        match f {
-            FileName::Internal(..) => true,
-            FileName::Custom(s) => s.starts_with('<'),
-            _ => false,
-        }
-    }
-}
-
-pub(crate) fn minify_file_comments(
-    comments: &SingleThreadedComments,
-    preserve_comments: BoolOr<JsMinifyCommentOption>,
-) {
-    match preserve_comments {
-        BoolOr::Bool(true) | BoolOr::Data(JsMinifyCommentOption::PreserveAllComments) => {}
-
-        BoolOr::Data(JsMinifyCommentOption::PreserveSomeComments) => {
-            let preserve_excl = |_: &BytePos, vc: &mut Vec<Comment>| -> bool {
-                // Preserve license comments.
-                //
-                // See https://github.com/terser/terser/blob/798135e04baddd94fea403cfaab4ba8b22b1b524/lib/output.js#L175-L181
-                vc.retain(|c: &Comment| {
-                    c.text.contains("@lic")
-                        || c.text.contains("@preserve")
-                        || c.text.contains("@copyright")
-                        || c.text.contains("@cc_on")
-                        || (c.kind == CommentKind::Block && c.text.starts_with('!'))
-                });
-                !vc.is_empty()
-            };
-            let (mut l, mut t) = comments.borrow_all_mut();
-
-            l.retain(preserve_excl);
-            t.retain(preserve_excl);
-        }
-
-        BoolOr::Bool(false) => {
-            let (mut l, mut t) = comments.borrow_all_mut();
-            l.clear();
-            t.clear();
-        }
-    }
-}
-
 /// High-level apis.
 impl Compiler {
     pub fn new(cm: Arc<SourceMap>) -> Self {
